Shipment Service Changes

Related work items: #97653, #101099, #107699, #110760, #110882, #111241, #111265, #111284, #111312, #112078, #112104, #113199, #113784, #113796, #113885, #114654, #115523, #116757, #116975, #116999, #117555, #117920, #119427, #119650, #120066, #120088, #120858, #121366, #121741, #121755, #122014, #122124, #122164, #122182, #122186, #122903

+++ src/main/java/com/dpw/runner/shipment/services/dto/request/awb/AwbOCIInfo.java
@@ -12,7 +12,7 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AwbOCIInfo {
-    private Integer entityId;
+    private Long entityId;
     private String entityType;
     private Integer informationIdentifier;
     private Integer tradeIdentificationCode;
